chore: choices ux improved and implemented useProcessDate hook
